@@ -3,8 +3,8 @@ package io.bidmachine.qr_ad_fetcher.activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
-import androidx.appcompat.app.AppCompatActivity
 import io.bidmachine.qr_ad_fetcher.AdType
 import io.bidmachine.qr_ad_fetcher.Helper
 import io.bidmachine.qr_ad_fetcher.NetworkRequest
@@ -13,10 +13,11 @@ import io.bidmachine.qr_ad_fetcher.ad.Ad
 import io.bidmachine.qr_ad_fetcher.ad.BannerAd
 import io.bidmachine.qr_ad_fetcher.ad.InterstitialAd
 import io.bidmachine.qr_ad_fetcher.ad.VideoAd
-import kotlinx.android.synthetic.main.activity_ad_manager.*
+import io.bidmachine.qr_ad_fetcher.databinding.ActivityAdManagerBinding
 import java.net.URL
 
-class AdManagerActivity : AppCompatActivity(), NetworkRequest.Listener, Ad.Listener {
+class AdManagerActivity
+    : BindingActivity<ActivityAdManagerBinding>(), NetworkRequest.Listener, Ad.Listener {
 
     companion object {
 
@@ -35,29 +36,32 @@ class AdManagerActivity : AppCompatActivity(), NetworkRequest.Listener, Ad.Liste
     private lateinit var adType: AdType
     private var ad: Ad? = null
 
+    override fun inflate(inflater: LayoutInflater) = ActivityAdManagerBinding.inflate(inflater)
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_ad_manager)
 
         adType = intent.getSerializableExtra(BUNDLE_AD_TYPE) as AdType
         val url = intent.getSerializableExtra(BUNDLE_URL) as URL
 
-        toolbar.setNavigationOnClickListener {
-            onBackPressed()
-            finish()
-        }
-        toolbar.setOnMenuItemClickListener {
-            if (it.itemId == R.id.home) {
-                val intent = MainActivity.getNewIntent(this)
-                startActivity(intent)
-                return@setOnMenuItemClickListener true
+        binding.toolbar.apply {
+            setNavigationOnClickListener {
+                onBackPressed()
+                finish()
             }
-            return@setOnMenuItemClickListener false
+            setOnMenuItemClickListener {
+                if (it.itemId == R.id.home) {
+                    val intent = MainActivity.getNewIntent(this@AdManagerActivity)
+                    startActivity(intent)
+                    return@setOnMenuItemClickListener true
+                }
+                return@setOnMenuItemClickListener false
+            }
         }
-        bLoad.setOnClickListener {
+        binding.bLoad.setOnClickListener {
             loadAd(url)
         }
-        bShow.setOnClickListener {
+        binding.bShow.setOnClickListener {
             showAd()
         }
     }
@@ -69,15 +73,15 @@ class AdManagerActivity : AppCompatActivity(), NetworkRequest.Listener, Ad.Liste
     }
 
     private fun loadAd(url: URL) {
-        bShow.isEnabled = false
-        progressBar.visibility = View.VISIBLE
+        binding.bShow.isEnabled = false
+        binding.progressBar.visibility = View.VISIBLE
         destroyAd()
         NetworkRequest.getBodyByUrl(url, this)
     }
 
     private fun prepareAd(adType: AdType, adm: String) {
         when (adType) {
-            AdType.Banner -> BannerAd(this, adContainer)
+            AdType.Banner -> BannerAd(this, binding.adContainer)
             AdType.Interstitial -> InterstitialAd(this)
             AdType.Video -> VideoAd(this)
         }.apply {
@@ -105,20 +109,20 @@ class AdManagerActivity : AppCompatActivity(), NetworkRequest.Listener, Ad.Liste
     }
 
     override fun onError() {
-        bShow.isEnabled = false
-        progressBar.visibility = View.INVISIBLE
+        binding.bShow.isEnabled = false
+        binding.progressBar.visibility = View.INVISIBLE
         Helper.showToast(this, "Can't load creative by url")
     }
 
     override fun onAdLoaded() {
-        bShow.isEnabled = true
-        progressBar.visibility = View.INVISIBLE
+        binding.bShow.isEnabled = true
+        binding.progressBar.visibility = View.INVISIBLE
         Helper.showToast(this, "Ad Loaded")
     }
 
     override fun onAdFailedToLoad() {
-        bShow.isEnabled = false
-        progressBar.visibility = View.INVISIBLE
+        binding.bShow.isEnabled = false
+        binding.progressBar.visibility = View.INVISIBLE
         Helper.showToast(this, "Ad Failed To Load")
     }
 
