@@ -1,22 +1,31 @@
 package io.bidmachine.qr_ad_fetcher.ad
 
+import android.app.Activity
 import android.content.Context
-import com.explorestack.iab.utils.Utils
-import io.bidmachine.qr_ad_fetcher.Helper
 
 interface Ad {
 
     fun loadAd(context: Context, adm: String)
-    fun showAd(context: Context)
+
+    fun showAd(activity: Activity)
+
     fun destroy()
 
+
     interface Listener {
+
         fun onAdLoaded()
+
         fun onAdFailedToLoad()
+
         fun onAdShown()
+
         fun onAdFailedToShown()
+
         fun onAdClicked()
+
         fun onAdClosed()
+
     }
 
 }
